Remove junit-jupiter-api to get junit 4 tests to run
If the junit-jupiter-api dependency is declared and the
project contains junit 4 tests, we would have to add
the junit jupyter vintage engine as a dependency as well
to get the tests running.
As long as the vintage engine is missing, the junit 4
tests are not executed!
If the project only contains junit4 tests, a simple solution
is to remove junit-jupiter-api. In this case classic junit
4 is used and the tests are executed.