Sample and Documentation: IAS identity kernel service (#272)
This sample is a Java back-end application running on the Cloud Foundry. For all incoming requests it checks whether the IAS token is valid.
Technically, the user is authenticated using the IasTokenAuthenticator which is defined in the Java Security library.

The identity service is released with name "identity" and plan"application".