Improve the custom SecurityContext Strategy registration
The library now registers a security context strategy by default. In CAP Java we ran into situations where this registration happened too late.
The effect of this was, that SecurityConfigs used an obsolete SecurityContextStrategy (which stored by Spring Security as a member variable in the SecurityConfig) and the library set a new strategy in the static holder class.
As the SecurityConfigs stored the authentication objects in the obsolete strategy, CAP wasn't able anymore to get the current authentication from the static holder class.

This PR tries to avoid this situation, by ensuring that the auto-configuration has a high precedence and is run early in the application initialization.
Additional this PR avoids reflection usage to set the strategy, as reflection is not really required.