Remove dependency-check-maven plugin from build
Dependency check is done by dependabot already and does not need to
be done in the build. Furthermore, the build was unstable because
of connection issues while downloading vulnerability data from NVD.
Instead of investing into stabilizing the NVD based dependency
check (e.g. by caching the NVD data), the decision was to replace
it with a dependabot only approach.

As a consequence, during the release of a new library version, we
need to double check for any waiting dependabot PRs that might fix
a known vulnerability.