fix null value (and key) support in configuration builder (#1747)
A recent fix of the oauth configuration builder based on Map.copy for
copying the properties of an configuration object broke consumers
having null values (or keys) in the properties map of the configuration.

This change swaps tha call to Map.copy() with calling the constructor
of HashMap which supports null. Same is done for copying the list of
domains.

Fixes #1743